+++ src/utility/makeReq.ts
@@ -201,6 +201,19 @@ const makeRequest = async (url: string, args: RequestInit) => {
                 }
             } catch (err) {
                 counter++;
+                // BUG: https://github.com/nodejs/node/issues/47246
+                // if the header content is too large, it will throw an error like
+                // code: `UND_ERR_HEADERS_OVERFLOW`
+                // so, if this error happens, tell the user to fix it using setting the environment variables
+                // if this is docker, this will be increased by default
+                if (err.cause && err.cause.code === "UND_ERR_HEADERS_OVERFLOW") {
+                    console.log(
+                        chalk.yellow(
+                            `[!] The tool detected a header overflow. Please increase the limit by setting environment variable \`NODE_OPTIONS="--max-http-header-size=99999999"\`. If the error still persists, please try again with a higher limit.`
+                        )
+                    );
+                    process.exit(21);
+                }
                 if (counter > 10) {
                     console.log(chalk.red(`[!] Failed to fetch ${url} : ${err}`));
                     return null;