+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix errors when the target is using a CDN
 - Fix "possible EventEmitter memory leak detected" error
 - Fix "Maximum call stack size exceeded" error in resolveNodeValue
+- Fix "UND_ERR_HEADERS_OVERFLOW" error
 
 ## 1.2.1-alpha.2 - 2025.08.06